nov 20 steps finished

+++ api/Program.cs
@@ -16,7 +16,7 @@ app.UseCors(c =>
 app.MapGet("/", () => "Hello World!");
 
 var storageRoot = "./storage";
-var hashSetRoot = "./storage/hash.json";
+var hashSetRoot = "./hash";
 
 if (!Directory.Exists(storageRoot)) Directory.CreateDirectory(storageRoot);
 
@@ -24,31 +24,37 @@ if (!Directory.Exists(storageRoot)) Directory.CreateDirectory(storageRoot);
 // var myHashSetThing = File.ReadAllText("./storage/hashset.json");
 //   var thisIsATest = JsonSerializer.Serialize(ExistingUsers);
 //   File.WriteAllText("./storage/hashset.json", thisIsATest);
-  HashSet<string> ExistingUsers = [];
+HashSet<string> ExistingUsers = [];
 if (!Directory.Exists(hashSetRoot))
 {
   ExistingUsers = [];
+  Directory.CreateDirectory(hashSetRoot);
 }
 else
 {
-  var myHashSetThing = File.ReadAllText("./storage/hash.json");
+  var myHashSetThing = File.ReadAllText(hashSetRoot + "/hashObj.json");
   ExistingUsers = JsonSerializer.Deserialize<HashSet<string>>(myHashSetThing);
 }
+
 app.MapPost("/newUser", async (User user) =>
 {
-  ExistingUsers.Add(user.UserName);
-  var userPath = Path.Combine(storageRoot, Convert.ToString(user.UserName));
-  Directory.CreateDirectory(userPath);
+  if (!ExistingUsers.Contains(user.UserName))
+  {
+    ExistingUsers.Add(user.UserName);
+    File.WriteAllText(hashSetRoot + "/hashObj.json", JsonSerializer.Serialize(ExistingUsers));
+    var userPath = Path.Combine(storageRoot, Convert.ToString(user.UserName));
+    Directory.CreateDirectory(userPath);
 
-  var userFile = Path.Combine(userPath, "user.json");
-  await File.WriteAllTextAsync(userFile, JsonSerializer.Serialize(user));
+    var userFile = Path.Combine(userPath, "user.json");
+    await File.WriteAllTextAsync(userFile, JsonSerializer.Serialize(user));
+  }
 });
 
 app.MapGet("/userList", () =>
 {
   string[]? myUsers = new string[ExistingUsers.Count()];
   ExistingUsers.CopyTo(myUsers);
-  return JsonSerializer.Serialize(myUsers);
+  return myUsers;
 });
 
 app.MapGet("/user/{userName}", (string userName) =>
@@ -73,14 +79,24 @@ app.MapDelete("/user/{userName}/delete", (string userName) =>
     throw new Exception("User not found");
   }
   var blogFolder = userName;
-  var blogFile = Path.Combine(storageRoot, Convert.ToString(userName));
-  File.Delete(blogFile);
+  var blogFile = Path.Combine(storageRoot, userName);
+  ExistingUsers.Remove(userName);
+  File.WriteAllText(hashSetRoot + "/hashObj.json", JsonSerializer.Serialize(ExistingUsers));
+  Directory.Delete(blogFile, true);
 });
-// var blogPath = Path.Combine(storageRoot, blog.Id.ToString());
-// Directory.CreateDirectory(blogPath);
 
-// var blogFile = Path.Combine(blogPath, "blog.json");
-// await File.WriteAllTextAsync(blogFile, JsonSerializer.Serialize(blog));
+// app.MapGet("/blogs/{blogId}/comments", (ulong blogId) =>
+// {
+//   var commentsPath = Path.Combine(storageRoot, blogId.ToString(), "comments");
+//   if (!Directory.Exists(commentsPath))
+//     return [];
+
+//   var comments = Directory.GetFiles(commentsPath)
+//       .Select((file) => File.ReadAllText(file))
+//       .Select((rawComment) => JsonSerializer.Deserialize<Comment>(rawComment));
+
+//   return comments;
+// });
 
 app.Run();
